protocol: restjson: add `__type` field to error response struct
`__type` is a part of an error struct used by some services supported
rest json as a protocol (e.g., ks, ps). It contains an error code.
Another option for this code is the `code` field.

Now if this field is present in the error response, it will be used as a
 code in the resulting error.